Add Makefile to allow regenerating README.html and example.html.
I added this because it was nontrivial for me to figure out how to
regenerate these files as they were.

The dependencies are obviously insufficient, but I still believe it is
useful, at least to document how to regenerate these files as they are.
(I am also somewhat bothered by the way sub and toc interact, and that
the order they run changes things, and that the two generated files were
generated with their interaction in the opposite order.)

This leads to one substantive change in example.html that I could not
suppress: the DOCTYPE is preserved.  (Maybe this is an html5lib change?)

I intentionally reverted the update to the date, since there was no
change to the content.